Leave application status section complete.

+++ ums-api/src/main/java/org/ums/domain/model/immutable/common/LmsAppStatus.java
@@ -13,7 +13,7 @@ import java.util.Date;
 /**
  * Created by Monjur-E-Morshed on 06-May-17.
  */
-public interface LmsAppStatus extends Serializable, LastModifier, EditType<MutableLmsAppStatus>, Identifier<Integer> {
+public interface LmsAppStatus extends Serializable, LastModifier, EditType<MutableLmsAppStatus>, Identifier<Long> {
   Long getLmsAppId();
 
   LmsApplication getLmsApplication();
